feat(tables): replace Double precision and Bigint for Decimal(10,2)

+++ src/main/resources/db/changelog/changes/20250814-1600-initial-schema.sql
@@ -12,8 +12,8 @@ CREATE TABLE meal
     id         BIGSERIAL PRIMARY KEY,
     name       VARCHAR(50)      NOT NULL,
     quantity   INT              NOT NULL,
-    price      DOUBLE PRECISION NOT NULL,
-    image_url  VARCHAR(255)      NOT NULL,
+    price      DECIMAL(10, 2)   NOT NULL,
+    image_url  VARCHAR(255)     NOT NULL,
     CONSTRAINT uq_meal_name UNIQUE (name)
 );
 
@@ -31,7 +31,7 @@ CREATE TABLE "order"
 (
     id           BIGSERIAL PRIMARY KEY,
     status       VARCHAR(50) NOT NULL,
-    total_amount BIGINT      NOT NULL,
+    total_amount DECIMAL(10, 2)      NOT NULL,
     member_id    BIGINT      NOT NULL REFERENCES member (id),
     created_at   TIMESTAMP   NOT NULL,
     updated_at   TIMESTAMP   NOT NULL
@@ -39,18 +39,18 @@ CREATE TABLE "order"
 
 CREATE TABLE order_item
 (
-    id         BIGSERIAL PRIMARY KEY,
-    order_id   BIGINT           NOT NULL REFERENCES "order" (id) ON DELETE CASCADE,
-    meal_id  BIGINT           NOT NULL REFERENCES meal (id),
-    quantity   INT              NOT NULL,
-    price DOUBLE PRECISION NOT NULL
+    id          BIGSERIAL PRIMARY KEY,
+    order_id    BIGINT           NOT NULL REFERENCES "order" (id) ON DELETE CASCADE,
+    meal_id     BIGINT           NOT NULL REFERENCES meal (id),
+    quantity    INT              NOT NULL,
+    price       DECIMAL(10, 2)   NOT NULL
 );
 
 CREATE TABLE payment
 (
     id                       BIGSERIAL PRIMARY KEY,
     stripe_payment_intent_id VARCHAR(255) NOT NULL UNIQUE,
-    amount                   BIGINT       NOT NULL,
+    amount                   DECIMAL(10, 2)       NOT NULL,
     currency                 VARCHAR(50)  NOT NULL,
     status                   VARCHAR(50)  NOT NULL,
     failure_code             VARCHAR(255),
